feat: simplify auth flow
