+++ apps/web/app/api/auth/[...nextauth]/route.ts
@@ -1,17 +1,3 @@
-import NextAuth from 'next-auth';
-import { nextAuthOptions } from 'app/auth/options';
-import { NextApiRequest, NextApiResponse } from 'next';
-import { NextRequest, NextResponse } from 'next/server';
+import { handlers } from "app/auth/auth";
 
-async function handler(req: NextRequest, res: NextResponse) {
-  // Extract cookies from the request
-  const cookies = req.cookies;
-
-  // Pass the cookies or other request-specific data to nextAuthOptions
-  const options = nextAuthOptions(cookies);
-
-  // Create and return the NextAuth handler with the dynamically generated options
-  return await NextAuth(options)(req, res);
-}
-
-export { handler as GET, handler as POST };
+export const { GET, POST } = handlers
