+++ apps/web/app/auth/auth.ts
@@ -1,0 +1,4 @@
+import NextAuth from "next-auth"
+import authConfig from "./auth.config"
+ 
+export const { auth, handlers, signIn, signOut } = NextAuth(authConfig)
