+++ apps/web/app/layout.tsx
@@ -4,7 +4,7 @@ import StyledComponentsRegistry from '../components/Utils/libs/styled-registry'
 import { motion } from 'framer-motion'
 import { SessionProvider } from 'next-auth/react'
 import LHSessionProvider from '@components/Contexts/LHSessionContext'
-import { isDevEnv } from './auth/options'
+import { isDevEnv } from './auth/auth.config'
 import Script from 'next/script'
 import CookiesProvider from '@components/Contexts/CookiesContext'
