+++ apps/web/app/orgs/[orgslug]/login/login.tsx
@@ -68,7 +68,7 @@ const LoginClient = (props: LoginClientProps) => {
         redirect: false,
         email: values.email,
         password: values.password,
-        callbackUrl: '/redirect_from_auth'
+        callbackUrl: getUriWithOrg(props.org?.slug, '/')
       });
       if (res && res.error) {
         setError("Wrong Email or password");
@@ -77,7 +77,7 @@ const LoginClient = (props: LoginClientProps) => {
         await signIn('credentials', {
           email: values.email,
           password: values.password,
-          callbackUrl: '/redirect_from_auth'
+          callbackUrl: getUriWithOrg(props.org?.slug, '/')
         });
       }
     },
@@ -194,7 +194,7 @@ const LoginClient = (props: LoginClientProps) => {
               <UserRoundPlus size={17} />
               <span>Sign up</span>
             </Link>
-            <button onClick={() => signIn('google', { callbackUrl: '/redirect_from_auth' })} className="flex justify-center py-3 text-md w-full bg-white text-slate-600 space-x-3 font-semibold text-center p-2 rounded-md shadow hover:cursor-pointer">
+            <button onClick={() => signIn('google', { callbackUrl: getUriWithOrg(props.org?.slug, '/') })} className="flex justify-center py-3 text-md w-full bg-white text-slate-600 space-x-3 font-semibold text-center p-2 rounded-md shadow hover:cursor-pointer">
               <img src="https://fonts.gstatic.com/s/i/productlogos/googleg/v6/24px.svg" alt="" />
               <span>Sign in with Google</span>
             </button>
