+++ apps/web/app/orgs/[orgslug]/signup/page.tsx
@@ -12,7 +12,7 @@ type MetadataProps = {
 export async function generateMetadata(
   params
     : MetadataProps): Promise<Metadata> {
-  const orgslug = params.searchParams.orgslug
+  const orgslug = params.params.orgslug
   // Get Org context information
   const org = await getOrganizationContextInfo(orgslug, {
     revalidate: 0,
@@ -25,7 +25,7 @@ export async function generateMetadata(
 }
 
 const SignUp = async (params: any) => {
-  const orgslug = params.searchParams.orgslug
+  const orgslug = params.params.orgslug
   const org = await getOrganizationContextInfo(orgslug, {
     revalidate: 0,
     tags: ['organizations'],
