+++ apps/web/package.json
@@ -30,6 +30,7 @@
     "@tiptap/react": "^2.6.4",
     "@tiptap/starter-kit": "^2.6.4",
     "@types/js-cookie": "^3.0.6",
+    "@types/next-auth": "^3.15.0",
     "@types/randomcolor": "^0.5.9",
     "@upstash/redis": "^1.34.0",
     "avvvatars-react": "^0.4.2",
@@ -42,8 +43,8 @@
     "katex": "^0.16.11",
     "lowlight": "^3.1.0",
     "lucide-react": "^0.424.0",
-    "next": "14.2.5",
-    "next-auth": "^4.24.7",
+    "next": "14.2.7",
+    "next-auth": "5.0.0-beta.20",
     "nextjs-toploader": "^1.6.12",
     "prosemirror-state": "^1.4.3",
     "randomcolor": "^0.6.2",
