+++ apps/web/tsconfig.json
@@ -28,7 +28,12 @@
       "@editor/*": ["components/Objects/Editor/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/**/*.tsx", ".next/types/**/*.ts"],
-
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/**/*.tsx",
+    ".next/types/*.ts",
+    ".next/types/**/*.ts"
+  ],
   "exclude": ["node_modules"]
 }
